Refactor & simplify ConfigContext, Rename InitialPage => FirstPage
It's nice to have ConfigContext be solely responsible for lazypages.
It's also much simpler! So simple a builder is no longer necessary.
Ideally only ConfigContext would have access to FirstPage's sqliteheader
and sqliteschema. I'll have to think on that.
We could pull reading that info out of FirstPage and put it into
ConfigContext... We'll see.